+++ app/src/main/java/com/example/passwordgenerator/domain/repository/PasswordRepository.kt
@@ -10,7 +10,7 @@ interface PasswordRepository {
 
     fun getPasswordsByFolder(folderId: Int): LiveData<List<Password>>
 
-    suspend fun savePassword(password: Password): Int
+    suspend fun savePassword(password: Password): Long
 
     suspend fun deletePassword(passwordId: Int)
